Activate seal epoch after connecting tip
